Update default logger to be info
The default log-level of "warn" does not help much
even when there are issues with nlk:
- While talking to the API.
- updating upstreams.

This commit updates the default loglevel to be
info so that the customer sees useful logs while
debugging the component. The user can choose to
turn it down if they want to but having a slightly
louder log level helps with the initial experience
especially when the product is new (ish).